pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-aggregator"
version = "0.1.0"
description = "Plugin-based Telegram job aggregator that publishes normalized jobs to Kafka"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Telethon>=1.38,<2",
  "kafka-python>=2.0.2,<3",
  "pydantic>=2.8,<3",
  "PyYAML>=6.0,<7"
]

[project.scripts]
job-aggregator = "job_aggregator.app:cli"

[tool.setuptools.packages.find]
include = ["job_aggregator*"]
